Helper integration and small cleanup
